gitops/apps/sietch-scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium==4.17.2
webdriver_manager==4.0.1
Flask==3.0.2 